+++ paper/paper.bib
@@ -713,7 +713,7 @@ archivePrefix = {arXiv},
   Series                   = {Proceedings of the 9th Python in Science Conference (SciPy)},
   volume                   = {445},
   pages                    = {51-56},
-  doi                      = {10.0.25080/majora-92bf1922-00a},
+  doi                      = {10.25080/Majora-92bf1922-00a},
 }
 
 @InProceedings{McKinney:2011,
