### 2020-10-24 #2772 @danehkar
Invalid DOI was fixed.
