### 2020-10-28 #2772 @danehkar
I am fine with Addendum.

Please include that its dependent Python packages are different than its IDL version:

`AtomNeb` uses the FITS handling routines of the Python package `Astropy` [@Astropy:2013; @Astropy:2018] to retrieve the atomic data from its FITS files. It also requires the Python packages `NumPy` [@Walt:2011; @Harris:2020] and `pandas` [@McKinney:2010; @McKinney:2011; @McKinney:2017]. Its latest version can be installed directly from its repository on the GitHub sing ``python setup.py install``, and the stable version from the Python Package Index (PyPi) via ``pip install atomneb`` or alternatively from the Conda Python package manager via ``conda install -c conda-forge atomneb``. The online documentation, tutorials and examples are provided on the GitHub platform (https://github.com/atomneb/AtomNeb-py) and the Read the Docs documentation host (https://atomneb-py.readthedocs.io/).

@ARTICLE{Astropy:2013,
       author = {{Astropy Collaboration} and {Robitaille}, T. P. and
         {Tollerud}, E. J. and {Greenfield}, P. and {Droettboom}, M. and
         {Bray}, Erik and {Aldcroft}, Tom and {Davis}, Matt and
         {Ginsburg}, Adam and {Price-Whelan}, Adrian M. and
         {Kerzendorf}, Wolfgang E. and {Conley}, Alexander and {Crighton}, Neil and
         {Barbary}, Kyle and {Muna}, Demitri and {Ferguson}, Henry and
         {Grollier}, Fr{\'e}d{\'e}ric and {Parikh}, Madhura M. and
         {Nair}, Prasanth H. and {Unther}, Hans M. and {Deil}, Christoph and
         {Woillez}, Julien and {Conseil}, Simon and {Kramer}, Roban and
         {Turner}, James E.~H. and {Singer}, Leo and {Fox}, Ryan and
         {Weaver}, Benjamin A. and {Zabalza}, Victor and {Edwards}, Zachary I. and
         {Azalee Bostroem}, K. and {Burke}, D.~J. and {Casey}, Andrew R. and
         {Crawford}, Steven M. and {Dencheva}, Nadia and {Ely}, Justin and
         {Jenness}, Tim and {Labrie}, Kathleen and {Lim}, Pey Lian and
         {Pierfederici}, Francesco and {Pontzen}, Andrew and {Ptak}, Andy and
         {Refsdal}, Brian and {Servillat}, Mathieu and {Streicher}, Ole},
        title = "{Astropy: A community Python package for astronomy}",
      journal = {Astronomy and Astrophysics},
         year = 2013,
       volume = {558},
          eid = {A33},
        pages = {A33},
          doi = {10.1051/0004-6361/201322068},
archivePrefix = {arXiv},
       eprint = {1307.6212},
 primaryClass = {astro-ph.IM},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2013A&A...558A..33A},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}

@ARTICLE{Astropy:2018,
       author = {{Astropy Collaboration} and {Price-Whelan}, A.~M. and
         {Sip{\H{o}}cz}, B.~M. and {G{\"u}nther}, H.~M. and {Lim}, P.~L. and
         {Crawford}, S.~M. and {Conseil}, S. and {Shupe}, D.~L. and
         {Craig}, M.~W. and {Dencheva}, N. and {Ginsburg}, A. and {Vand
        erPlas}, J.~T. and {Bradley}, L.~D. and {P{\'e}rez-Su{\'a}rez}, D. and
         {de Val-Borro}, M. and {Aldcroft}, T.~L. and {Cruz}, K.~L. and
         {Robitaille}, T.~P. and {Tollerud}, E.~J. and {Ardelean}, C. and
         {Babej}, T. and {Bach}, Y.~P. and {Bachetti}, M. and {Bakanov}, A.~V. and
         {Bamford}, S.~P. and {Barentsen}, G. and {Barmby}, P. and
         {Baumbach}, A. and {Berry}, K.~L. and {Biscani}, F. and {Boquien}, M. and
         {Bostroem}, K.~A. and {Bouma}, L.~G. and {Brammer}, G.~B. and
         {Bray}, E.~M. and {Breytenbach}, H. and {Buddelmeijer}, H. and
         {Burke}, D.~J. and {Calderone}, G. and {Cano Rodr{\'\i}guez}, J.~L. and
         {Cara}, M. and {Cardoso}, J.~V.~M. and {Cheedella}, S. and {Copin}, Y. and
         {Corrales}, L. and {Crichton}, D. and {D'Avella}, D. and {Deil}, C. and
         {Depagne}, {\'E}. and {Dietrich}, J.~P. and {Donath}, A. and
         {Droettboom}, M. and {Earl}, N. and {Erben}, T. and {Fabbro}, S. and
         {Ferreira}, L.~A. and {Finethy}, T. and {Fox}, R.~T. and
         {Garrison}, L.~H. and {Gibbons}, S.~L.~J. and {Goldstein}, D.~A. and
         {Gommers}, R. and {Greco}, J.~P. and {Greenfield}, P. and
         {Groener}, A.~M. and {Grollier}, F. and {Hagen}, A. and {Hirst}, P. and
         {Homeier}, D. and {Horton}, A.~J. and {Hosseinzadeh}, G. and {Hu}, L. and
         {Hunkeler}, J.~S. and {Ivezi{\'c}}, {\v{Z}}. and {Jain}, A. and
         {Jenness}, T. and {Kanarek}, G. and {Kendrew}, S. and {Kern}, N.~S. and
         {Kerzendorf}, W.~E. and {Khvalko}, A. and {King}, J. and {Kirkby}, D. and
         {Kulkarni}, A.~M. and {Kumar}, A. and {Lee}, A. and {Lenz}, D. and
         {Littlefair}, S.~P. and {Ma}, Z. and {Macleod}, D.~M. and
         {Mastropietro}, M. and {McCully}, C. and {Montagnac}, S. and
         {Morris}, B.~M. and {Mueller}, M. and {Mumford}, S.~J. and {Muna}, D. and
         {Murphy}, N.~A. and {Nelson}, S. and {Nguyen}, G.~H. and
         {Ninan}, J.~P. and {N{\"o}the}, M. and {Ogaz}, S. and {Oh}, S. and
         {Parejko}, J.~K. and {Parley}, N. and {Pascual}, S. and {Patil}, R. and
         {Patil}, A.~A. and {Plunkett}, A.~L. and {Prochaska}, J.~X. and
         {Rastogi}, T. and {Reddy Janga}, V. and {Sabater}, J. and
         {Sakurikar}, P. and {Seifert}, M. and {Sherbert}, L.~E. and
         {Sherwood-Taylor}, H. and {Shih}, A.~Y. and {Sick}, J. and
         {Silbiger}, M.~T. and {Singanamalla}, S. and {Singer}, L.~P. and
         {Sladen}, P.~H. and {Sooley}, K.~A. and {Sornarajah}, S. and
         {Streicher}, O. and {Teuben}, P. and {Thomas}, S.~W. and
         {Tremblay}, G.~R. and {Turner}, J.~E.~H. and {Terr{\'o}n}, V. and
         {van Kerkwijk}, M.~H. and {de la Vega}, A. and {Watkins}, L.~L. and
         {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
         {Zabalza}, V. and {Astropy Contributors}},
        title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
      journal = {The Astronomical Journal},
         year = 2018,
        month = sep,
       volume = {156},
       number = {3},
          eid = {123},
        pages = {123},
          doi = {10.3847/1538-3881/aabc4f},
archivePrefix = {arXiv},
       eprint = {1801.02634},
 primaryClass = {astro-ph.IM},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2018AJ....156..123A},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}

@ARTICLE{Walt:2011,
       author = {{van der Walt}, S. and {Colbert}, S. C. and
         {Varoquaux}, G.},
        title = "{The NumPy Array: A Structure for Efficient Numerical Computation}",
      journal = {Computing in Science and Engineering},
     keywords = {Computer Science - Mathematical Software},
         year = 2011,
       volume = {13},
       number = {2},
        pages = {22-30},
          doi = {10.1109/MCSE.2011.37},
archivePrefix = {arXiv},
       eprint = {1102.1523},
 primaryClass = {cs.MS},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2011CSE....13b..22V},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}

@ARTICLE{Harris:2020,
       author = {{Harris}, Charles R. and {Jarrod Millman}, K. and
         {van der Walt}, S. J. and {Gommers}, R. and
         {Virtanen}, Pauli and {Cournapeau}, David and {Wieser}, Eric and
         {Taylor}, Julian and {Berg}, Sebastian and {Smith}, Nathaniel J. and
         {Kern}, Robert and {Picus}, Matti and {Hoyer}, Stephan and
         {van Kerkwijk}, Marten H. and {Brett}, Matthew and {Haldane}, Allan and
         {Fern{\'a}ndez del R{\'\i}o}, Jaime and {Wiebe}, Mark and
         {Peterson}, Pearu and {G{\'e}rard-Marchant}, Pierre and
         {Sheppard}, Kevin and {Reddy}, Tyler and {Weckesser}, Warren and
         {Abbasi}, Hameer and {Gohlke}, Christoph and {Oliphant}, Travis E.},
        title = "{Array Programming with NumPy}",
      journal = {Nature},
         year = 2020,
       volume = {585},
        pages = {357},
          doi = {10.1038/s41586-020-2649-2},
archivePrefix = {arXiv},
       eprint = {2006.10256},
 primaryClass = {cs.MS},
       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200610256H},
      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
}

@InProceedings{McKinney:2010,
  Title                    = {{Data Structures for Statistical Computing in Python}},
  Author                   = {{McKinney}, W.},
  Year                     = {2010},
  Series                   = {Proceedings of the 9th Python in Science Conference (SciPy)},
  volume                   = {445},
  pages                    = {51-56},
  doi                      = {10.25080/Majora-92bf1922-00a},
}

@InProceedings{McKinney:2011,
  Title                    = {{pandas: a Foundational Python Library for Data Analysis and Statistics}},
  Author                   = {{McKinney}, W.},
  Year                     = {2011},
  Series                   = {Python for High Performance and Scientific Computing},
  volume                   = {14},
  number                   = {9},
}

@Book{McKinney:2017,
  Title                    = {{Python for Data Analysis: Data Wrangling with Pandas, NumPy, and IPython, 2nd Edition}},
  Author                   = {{McKinney}, W.},
  isbn                     = {9781491957639},
  lccn                     = {2018302150},
  Year                     = {2017},
  Publisher                = {O'Reilly Media}
}
