### 2020-10-28 #2772 @danehkar
> After discussing with the JOSS editorial team, we've determined that https://github.com/atomneb/AtomNeb-py is out of scope for JOSS as it is a direct port of your earlier package (https://github.com/atomneb/AtomNeb-idl) and therefore does not reach our threshold for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort).
>
> While we can't support a pulication in JOSS related to this work, we would allow an update (addendum) to [your paper](https://joss.theoj.org/papers/10.21105/joss.00898) on `AtomNeb-idl` noting that there is now a Python port of this package.
>

@arfon, actually,  i think that a direct port is not a correct description. There is a possibility to port IDL into Python. However, the new python package is not a port of IDL version, it is purely a new implementation in Python language and translation of IDL version into Python code. You can see the Python codes here:

https://github.com/atomneb/AtomNeb-py/blob/master/atomneb/atomneb.py

Atomic data is stored in the **structure** in IDL version, while atomic data is stored using the **dictionary** and **pandas**python  library in Python. IDL version uses the IDL **Astronomy User's Library** (https://idlastro.gsfc.nasa.gov/homepage.html), while the Python version uses **AstroPy** (https://www.astropy.org/).

I think that the correct phrase is a Python implementation (not a Python port) of this package.
