### 2020-10-28 #2772 @danehkar
@arfon I am happy to update the published paper and mention its Python implementation. Can we also revise the title of my published paper? For example, to "AtomNeb: IDL/Python Package for Atomic Data of Ionized Nebulae"
