### 2020-10-28 #2772 @danehkar
> Similar to my [comment](https://github.com/openjournals/joss-reviews/issues/2773#issuecomment-718161693) on `pyEQUIB`, we can't modify the original paper but we can make sure the new package is mentioned in the addendum title, e.g.:
>
> `AtomNebpy, an addendum to AtomNeb: IDL Library for Atomic Data of Ionized Nebulae`
>
> (or something like that)

The name of python package is still atomneb:
https://pypi.org/project/atomneb/
https://pypi.org/project/atomneb/

so it is better to be `AtomNeb Python Package, an addendum to AtomNeb: IDL Library for Atomic Data of Ionized Nebulae`
