### 2020-10-28 #2772 @arfon
Sounds good to me @danehkar. I've stubbed out a basic addendum format [here](https://gist.githubusercontent.com/arfon/17449e908eaa77112a8027e91d7116a0/raw/ff72f34f4971b388acd56f0f8f90e736ed2eb0f5/paper.md). Could you update the `paper.md` in the https://github.com/atomneb/AtomNeb-py repository with a paragraph describing the Python package?

Also, please do something similar over in https://github.com/openjournals/joss-reviews/issues/2773 for that package too.
