### 2020-10-29 #2797 @whedon
[REVIEW]: AtomNeb: Python Package for Atomic Data of Ionized Nebulae
**Submitting author:** @danehkar (<a href="http://orcid.org/0000-0003-4552-5997">Ashkbiz Danehkar</a>)
**Repository:** <a href="https://github.com/atomneb/AtomNeb-py" target ="_blank">https://github.com/atomneb/AtomNeb-py</a>
**Version:** v0.2.3
