+++ paper/paper.bib
@@ -1,333 +1,3 @@
-@article{Landi:2006,
-   author = {{Landi}, E. and {Del Zanna}, G. and {Young}, P.~R. and {Dere}, K.~P. and 
-	{Mason}, H.~E. and {Landini}, M.},
-    title = "{CHIANTI-An Atomic Database for Emission Lines. VII. New Data for X-Rays and Other Improvements}",
-  journal = {The Astrophysical Journal Supplement Series},
-     year = 2006,
-   volume = 162,
-    pages = {261-280},
-      doi = {10.1086/498148},
-   adsurl = {http://adsabs.harvard.edu/abs/2006ApJS..162..261L},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{Dere:2009,
-   author = {{Dere}, K.~P. and {Landi}, E. and {Young}, P.~R. and {Del Zanna}, G. and 
-	{Landini}, M. and {Mason}, H.~E.},
-    title = "{CHIANTI - an atomic database for emission lines. IX. Ionization rates, recombination rates, ionization equilibria for the elements hydrogen through zinc and updated atomic data}",
-  journal = {Astronomy and Astrophysics},
-     year = 2009,
-   volume = 498,
-    pages = {915-929},
-      doi = {10.1051/0004-6361/200911712},
-   adsurl = {http://adsabs.harvard.edu/abs/2009A%26A...498..915D},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{Dere:2019,
-       author = {{Dere}, K.~P. and {Del Zanna}, G. and {Young}, P.~R. and {Landi}, E. and
-         {Sutherland}, R.~S.},
-        title = "{CHIANTI{\textemdash}An Atomic Database for Emission Lines. XV. Version 9, Improvements for the X-Ray Satellite Lines}",
-      journal = {The Astrophysical Journal Supplement Series},
-         year = 2019,
-       volume = {241},
-       number = {2},
-          eid = {22},
-        pages = {22},
-          doi = {10.3847/1538-4365/ab05cf},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ApJS..241...22D},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{Landi:2012,
-   author = {{Landi}, E. and {Del Zanna}, G. and {Young}, P.~R. and {Dere}, K.~P. and 
-	{Mason}, H.~E.},
-    title = "{CHIANTI{\mdash}An Atomic Database for Emission Lines. XII. Version 7 of the Database}",
-  journal = {The Astrophysical Journal},
-     year = 2012,
-   volume = 744,
-      eid = {99},
-    pages = {99},
-      doi = {10.1088/0004-637X/744/2/99},
-   adsurl = {http://adsabs.harvard.edu/abs/2012ApJ...744...99L},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{Luridiana:2015,
-   author = {{Luridiana}, V. and {Morisset}, C. and {Shaw}, R.~A.},
-    title = "{PyNeb: a new tool for analyzing emission lines. I. Code description and validation of results}",
-  journal = {Astronomy and Astrophysics},
-archivePrefix = "arXiv",
-   eprint = {1410.6662},
- primaryClass = "astro-ph.IM",
-     year = 2015,
-   volume = 573,
-      eid = {A42},
-    pages = {A42},
-      doi = {10.1051/0004-6361/201323152},
[…]
