+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "AtomNeb: Python Package for Atomic Data of Ionized Nebulae"
+title: "AtomNeb Python Package, an addendum to AtomNeb: IDL Library for Atomic Data of Ionized Nebulae"
 tags:
   - python
   - astrophysics
@@ -21,24 +21,11 @@ date: 20 October 2020
 bibliography: paper.bib
 ---
 
-# Introduction
+# Addendum
 
-Emission-line spectra emitted from gaseous nebulae can be used to probe the physical and chemical proprieties of the interstellar medium in our Galaxy and other galaxies. The collisional excitation and recombination processes in ionized gases produces collisionally excited lines (CEL) and recombination lines (RL), respectively. The spectral analysis of CELs can be performed by solving statistical equilibrium equations, which requires the energy levels ($E_{j}$), collision strengths ($\Omega_{ij}$), and transition probabilities ($A_{ij}$) of ions, and yields atomic level populations and emissivities. The results of these calculations are employed to determine physical conditions and chemical abundances from CELs [see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2014b]. Moreover, the emissivities of RLs can be determined from effective recombination coefficients ($\alpha_{\rm eff}$) of ions, which are utilized to estimate chemical abundances from RLs.
+`AtomNeb` is a Python open-source package containing atomic data for gaseous nebulae stored in the Flexible Image Transport System (FITS) file format [@Wells:1981; @Hanisch:2001; @Pence:2010]. These FITS files offer easy access to the atomic data required for emissivity calculations in the collisional excitation and recombination processes usually occurred in ionized gases of planetary nebulae and H II regions. This package has several application programming interface (API) functions developed in Python for retrieving the energy levels, collision strengths, transition probabilities, and recombination coefficients from its FITS files. The previous library `AtomNeb` [@Danehkar:2019] coupled to the library `proEQUIB` [@Danehkar:2018b] needs the Interactive Data Language (IDL) compiler, so this package offers an identical package for the high-level programming language Python that can be used by those astrophysicists, who intend to analyze nebular emission lines by developing codes in Python. Similarly, the `AtomNeb` Python functions can be used by the Python package `pyEQUIB` [@Danehkar:2020] to analyze emission-line spectra. 
 
-# Statement of Need
-
-The atomic data of the collisional excitation and recombination processes are essential for spectral analysis tools and photoionization programs. The collisional excitation calculations were implemented in some spectral analysis packages [e.g. @Howarth:1981; @Shaw:1994; @Shaw:1998; @Wesson:2012; @Luridiana:2015; @Howarth:2016; @Danehkar:2018b; @Danehkar:2020] and photoionization programs [e.g. @Ferland:1998; @Kallman:2001; @Ercolano:2003; @Ercolano:2005; @Ercolano:2008], which require to have the collision strengths ($\Omega_{ij}$) and transition probabilities ($A_{ij}$). Moreover, the recombination calculations in spectral analysis and photoionization codes [e.g. @Ferland:1998; @Kallman:2001; @Ercolano:2003; @Ercolano:2005; @Ercolano:2008; @Wesson:2012; @Luridiana:2015; @Danehkar:2018b; @Danehkar:2020] are performed using effective recombination coefficients ($\alpha_{\rm eff}$). The previous library `AtomNeb` [@Danehkar:2019] needs the Interactive Data Language (IDL) compiler, so it was necessary to develop an identical package for the high-level programming language Python that can be used by those astrophysicists, who intend to analyze nebular emission lines by developing codes in Python.
-
-# Description
-
[…]
