### 2020-11-23 #2797 @danehkar
@arfon I made a new release (v0.3.3) in its GitHub repository ([https://github.com/atomneb/AtomNeb-py](https://github.com/atomneb/AtomNeb-py)) and also an archive of the software in Zenodo service:

doi:[10.5281/zenodo.4287566](https://doi.org/10.5281/zenodo.4287566)

Thank you for reviewing this package.
